Deleted CallsFragment.kt and Moved DataAndStorageFragment.kt to storage package
